+++ src/device.rs
@@ -7,8 +7,8 @@ use crate::net::Ipv4Addr;
 use crate::socket::Socket;
 use crate::uninitialized_device::UninitializedDevice;
 use crate::{
-    common::{RetryCount, RetryTime},
-    register, MacAddress,
+    register::{self, common::RetryTime},
+    MacAddress, Mode,
 };
 
 pub enum ResetError<E> {
@@ -56,7 +56,7 @@ impl<SpiBus: Bus, HostImpl: Host> Device<SpiBus, HostImpl> {
 
     fn clear_mode(&mut self) -> Result<(), SpiBus::Error> {
         // Set RST common register of the w5500
-        self.as_mut().reset()
+        self.as_mut().reset_device()
     }
 
     #[inline]
@@ -119,21 +119,14 @@ impl<SpiBus: Bus, HostImpl: Host> Device<SpiBus, HostImpl> {
     }
 
     /// Get the current Retry Count Register value.
-    ///
-    /// RCR (Retry Count Register) [R/W] [0x001B] [0x08]
-    ///
-    /// E.g. In case of errors it will retry for 7 times:
-    /// `RCR = 0x0007`
     #[inline]
-    pub fn current_retry_count(&mut self) -> Result<RetryCount, SpiBus::Error> {
+    pub fn current_retry_count(&mut self) -> Result<u8, SpiBus::Error> {
         self.as_mut().current_retry_count()
     }
 
     /// Set a new value for the Retry Count register.
-    ///
-    /// RCR (Retry Count Register) [R/W] [0x001B] [0x08]
     #[inline]
-    pub fn set_retry_count(&mut self, retry_count: RetryCount) -> Result<(), SpiBus::Error> {
+    pub fn set_retry_count(&mut self, retry_count: u8) -> Result<(), SpiBus::Error> {
         self.as_mut().set_retry_count(retry_count)
     }
 
@@ -229,24 +222,25 @@ impl<SpiBus: Bus, HostImpl: Host> DeviceRefMut<'_, SpiBus, HostImpl> {
         Ok(phy[0].into())
     }
 
-    /// Reset the device
     #[inline]
-    fn reset(&mut self) -> Result<(), SpiBus::Error> {
-        self.set_mode(register::common::Mode::Reset)
+    pub fn reset_device(&mut self) -> Result<(), SpiBus::Error> {
+        // Set RST common register of the w5500
+        let mode = [0b10000000];
+        self.bus
+            .write_frame(register::COMMON, register::common::MODE, &mode)
     }
 
     #[inline]
-    fn set_mode(&mut self, mode_options: register::common::Mode) -> Result<(), SpiBus::Error> {
+    pub fn set_mode(&mut self, mode_options: Mode) -> Result<(), SpiBus::Error> {
         self.bus.write_frame(
             register::COMMON,
             register::common::MODE,
             &mode_options.to_register(),
-        )?;
-        Ok(())
+        )
     }
 
     #[inline]
-    fn version(&mut self) -> Result<u8, SpiBus::Error> {
+    pub fn version(&mut self) -> Result<u8, SpiBus::Error> {
         let mut version_register = [0_u8];
         self.bus.read_frame(
             register::COMMON,
@@ -272,7 +266,7 @@ impl<SpiBus: Bus, HostImpl: Host> DeviceRefMut<'_, SpiBus, HostImpl> {
     /// assert_eq!(four_hundred_ms.to_u16(), 4000);
     /// ```
     #[inline]
-    fn set_retry_timeout(&mut self, retry_time_value: RetryTime) -> Result<(), SpiBus::Error> {
+    pub fn set_retry_timeout(&mut self, retry_time_value: RetryTime) -> Result<(), SpiBus::Error> {
         self.bus.write_frame(
             register::COMMON,
             register::common::RETRY_TIME,
@@ -286,7 +280,7 @@ impl<SpiBus: Bus, HostImpl: Host> DeviceRefMut<'_, SpiBus, HostImpl> {
     ///
     /// E.g. 4000 = 400ms
     #[inline]
-    fn current_retry_timeout(&mut self) -> Result<RetryTime, SpiBus::Error> {
+    pub fn current_retry_timeout(&mut self) -> Result<RetryTime, SpiBus::Error> {
         let mut retry_time_register: [u8; 2] = [0, 0];
         self.bus.read_frame(
             register::COMMON,
@@ -300,23 +294,37 @@ impl<SpiBus: Bus, HostImpl: Host> DeviceRefMut<'_, SpiBus, HostImpl> {
     /// Set a new value for the Retry Count register.
     ///
     /// RCR (Retry Count Register) [R/W] [0x001B] [0x08]
-    fn set_retry_count(&mut self, retry_count: RetryCount) -> Result<(), SpiBus::Error> {
+    ///
+    /// For more details check out the rest of the datasheet documentation on the Retry count.
+    ///
+    /// From datasheet:
+    ///
+    /// RCR configures the number of time of retransmission. When retransmission occurs
+    /// as many as ‘RCR+1’, Timeout interrupt is issued (Sn_IR[TIMEOUT] = ‘1’).
+    ///
+    /// The timeout of W5500 can be configurable with RTR and RCR. W5500 has two kind
+    /// timeout such as Address Resolution Protocol (ARP) and TCP retransmission.
+    ///
+    /// E.g. In case of errors it will retry for 7 times:
+    /// `RCR = 0x0007`
+    pub fn set_retry_count(&mut self, retry_count: u8) -> Result<(), SpiBus::Error> {
         self.bus.write_frame(
             register::COMMON,
             register::common::RETRY_COUNT,
-            &retry_count.to_register(),
+            &[retry_count],
         )?;
 
         Ok(())
     }
 
     /// Get the current Retry Count value
+    ///
     /// RCR (Retry Count Register) [R/W] [0x001B] [0x08]
     ///
     /// E.g. In case of errors it will retry for 7 times:
     /// `RCR = 0x0007`
     #[inline]
-    fn current_retry_count(&mut self) -> Result<RetryCount, SpiBus::Error> {
+    pub fn current_retry_count(&mut self) -> Result<u8, SpiBus::Error> {
         let mut retry_count_register: [u8; 1] = [0];
         self.bus.read_frame(
             register::COMMON,
@@ -324,6 +332,6 @@ impl<SpiBus: Bus, HostImpl: Host> DeviceRefMut<'_, SpiBus, HostImpl> {
             &mut retry_count_register,
         )?;
 
-        Ok(RetryCount::from_register(retry_count_register))
+        Ok(retry_count_register[0])
     }
 }
